+++ src/main/kotlin/io/kjson/deserialize/KotlinFieldDescriptor.kt
@@ -2,7 +2,7 @@
  * @(#) KotlinFieldDescriptor.kt
  *
  * kjson  Reflection-based JSON serialization and deserialization for Kotlin
- * Copyright (c) 2024 Peter Wall
+ * Copyright (c) 2024, 2025 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -28,6 +28,9 @@ package io.kjson.deserialize
 import kotlin.reflect.KMutableProperty
 import kotlin.reflect.KProperty
 import kotlin.reflect.KType
+import io.kjson.JSONConfig
+import io.kjson.JSONDeserializer.applyTypeParameters
+import io.kjson.JSONDeserializer.findDeserializer
 
 data class KotlinFieldDescriptor<T>(
     override val propertyName: String,
@@ -46,6 +49,28 @@ data class KotlinFieldDescriptor<T>(
 
     override fun setValue(instance: Any, value: Any?) {
         (kProperty as KMutableProperty).setter.call(instance, value)
+    }
+
+    companion object {
+
+        fun create(
+            member: KProperty<*>,
+            resultType: KType,
+            config: JSONConfig,
+            references: MutableList<KType>,
+        ):  KotlinFieldDescriptor<*> {
+            val propertyName = config.findNameFromAnnotation(member.annotations) ?: member.name
+            val fieldType = member.getter.returnType.applyTypeParameters(resultType)
+            val deserializer = findDeserializer<Any>(fieldType, config, references)
+            return KotlinFieldDescriptor(
+                propertyName = propertyName,
+                ignore = config.hasIgnoreAnnotation(member.annotations),
+                nullable = fieldType.isMarkedNullable,
+                deserializer = deserializer,
+                kProperty = member,
+            )
+        }
+
     }
 
 }